refactored StatusView
